adjust test_bootstrapping_sets_up_tables & test_query_catalog_columns test re. ColId

+++ crates/core/src/vm.rs
@@ -779,7 +779,7 @@ pub(crate) mod tests {
             .with_select_cmp(
                 OpCmp::Eq,
                 FieldName::new(ST_COLUMN_ID, StColumnFields::ColPos.into()),
-                scalar(StColumnFields::TableId as u32),
+                scalar(StColumnFields::TableId as u16),
             )
             .unwrap();
         let st_column_row = StColumnRow {
